move commandHandler and contextGenerator files, and update import paths accordingly
